Move invitation generation logic into an independent component
